requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
Jinja2==2.11.2
psycopg2==2.8.5
python-dotenv==0.13.0
requests==2.23.0
SQLAlchemy==1.3.17
Flask-Migrate==2.5.3
